requirements.txt
pymongo
requests
python-dotenv
pandas
scikit-learn
joblib
streamlit
shap
matplotlib
xgboost
lightgbm
plotly
